+++ dflib/src/test/java/com/nhl/dflib/DataFrame_Sort_Test.java
@@ -32,6 +32,29 @@ public class DataFrame_Sort_Test {
     }
 
     @Test
+    public void testSort_NullsLast() {
+        DataFrame dfi = DataFrame.newFrame("a", "b").foldByRow(
+                0, 1,
+                null, 3,
+                -1, 2);
+
+        DataFrame df = dfi.sort(r -> (Integer) r.get("a"));
+        assertNotSame(dfi, df);
+
+        new DataFrameAsserts(dfi, "a", "b")
+                .expectHeight(3)
+                .expectRow(0, 0, 1)
+                .expectRow(1, null, 3)
+                .expectRow(2, -1, 2);
+
+        new DataFrameAsserts(df, "a", "b")
+                .expectHeight(3)
+                .expectRow(0, -1, 2)
+                .expectRow(1, 0, 1)
+                .expectRow(2, null, 3);
+    }
+
+    @Test
     public void testSortByColumns_Names() {
         DataFrame dfi = DataFrame.newFrame("a", "b").foldByRow(
                 0, 4,
@@ -55,7 +78,32 @@ public class DataFrame_Sort_Test {
                 .expectRow(0, 0, 2)
                 .expectRow(1, 2, 2)
                 .expectRow(2, 0, 4);
+    }
 
+    @Test
+    public void testSortByColumns_Names_Nulls() {
+        DataFrame dfi = DataFrame.newFrame("a", "b").foldByRow(
+                0, 4,
+                2, null,
+                0, 2);
+
+        DataFrame dfab = dfi.sort(new String[]{"a", "b"}, new boolean[]{true, true});
+        assertNotSame(dfi, dfab);
+
+        new DataFrameAsserts(dfab, "a", "b")
+                .expectHeight(3)
+                .expectRow(0, 0, 2)
+                .expectRow(1, 0, 4)
+                .expectRow(2, 2, null);
+
+        DataFrame dfba = dfi.sort(new String[]{"b", "a"}, new boolean[]{true, true});
+        assertNotSame(dfi, dfba);
+
+        new DataFrameAsserts(dfba, "a", "b")
+                .expectHeight(3)
+                .expectRow(0, 0, 2)
+                .expectRow(1, 0, 4)
+                .expectRow(2, 2, null);
     }
 
     @Test